Updated Deathnote domain and display
Moved some code from display to domain so that it can be used for Tome
